// d3jsthinpossample/src/Chart/Combo/Types.ts
export interface ComboCahartData {
    header: string;
    totalSum: number;
    totalUri: number;
    custumerSum: number;
}

export interface MaxValues {
    maxValue1: number;
    maxValue2: number;
    maxValue3: number;
}

export interface Size {
    width: number;
    height: number;
    marginTop: number;
    marginRight: number;
    marginBottom: number;
    marginLeft: number;
    //Y軸2-3の間
    margin2_3: number;
}

export interface Titele {
    title: string;
    yTitle1: string;
    yTitle2: string;
    yTitle3: string;
}

export interface Colors{
    data1Color: string;
    data2Color: string;
    data3Color: string;
}